requirements.txt
google-adk>=0.1.0
google-genai>=0.5.0
sympy>=1.12
numpy>=1.26.0
scipy>=1.13.0
pydantic>=2.0
python-dotenv>=1.0.0
rich>=13.0
langchain-community>=0.2.0
tavily-python>=0.3.0
crewai-tools>=0.3.0
